requirements.txt
django==4.2.11
django-filter==23.5
djangorestframework==3.14.0
django-widget-tweaks==1.5.0
psycopg2-binary==2.9.9
Pillow==10.2.0
django-tinymce==3.7.1
numpy==1.26.4
pandas==2.2.1
openpyxl==3.1.2
imap-tools==1.5.0
gunicorn==21.2.0
requests==2.31.0
apscheduler==3.10.4
sqlalchemy==2.0.28

flake8==7.0.0
model_bakery==1.17.0
drf_spectacular==0.27.1
django-silk==5.1.0